requirements.txt
# AR-Agent Requirements
# Medical Multimodal Augmented Reality Agent Dependencies

# Core ML/AI Libraries
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
accelerate>=0.21.0
peft>=0.4.0
bitsandbytes>=0.39.0
tokenizers>=0.15.0
sentencepiece>=0.1.99

# Web Framework
Flask>=2.3.0
requests>=2.31.0
httpx>=0.24.0
uvicorn>=0.23.0
fastapi>=0.100.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-image>=0.20.0
imageio>=2.28.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning Utilities
einops>=0.6.0
einops-exts>=0.0.4
timm>=0.9.0
safetensors>=0.3.0

# Data Processing
jsonlines>=3.1.0
markdown2>=2.4.0
pydantic>=2.0.0
shortuuid>=1.0.0

# Gradio for UI (optional)
gradio>=4.16.0
gradio-client>=0.8.0

# Medical Image Processing
SimpleITK>=2.2.0
pydicom>=2.4.0
nibabel>=5.1.0

# AR/3D Visualization
open3d>=0.17.0
vtk>=9.2.0
mayavi>=4.8.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
termcolor>=2.3.0
rich>=13.4.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Optional: Training Dependencies
# deepspeed>=0.12.6
# ninja>=1.11.0
# wandb>=0.15.0